+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/ ), and followin
 ### Added
  - added an option to teleport to a chunkloader (if you have the permission)
  - Added message when not able to load config. (#20)
+ - Added option to use another item instead of a blaze rod to init the chunk loaders.
 
 ### Changed
  -  the info command is now only showing active currently loaded chunks.